fix: clear lint-static-analysis CI break across all lanes
The lint-static-analysis lane fails before any tests run, which masks every
other lane's intent. Three independent breaks landed in the 2026-05-29/30
window after the last allowlist/baseline refresh:

- type-parity guard: migrate the fixed-point SILK pitch-lag scratch storage
  (silk_encode_frame_FIX and the find_pred_coefs/noise_shape/ltp drivers) from
  plain int to int32, matching libopus opus_int width and the existing pitchL
  convention; record the source-cited C double helpers (celt/custom mode window
  + log2 table, the arm64 MDCT FMA helper, and silk_sigmoid) in the allowlist.
- errcheck: discard the cleanup-path d.finishFixedHybrid() error explicitly on
  the hybrid decode-error branch.
- ineffassign: drop the dead masked TOC-byte assignment in the packet TOC edge
  parity test; libopus preserves the full TOC byte.
- revive unexported-return: scope-exclude the CELT parity trace hooks, which
  deliberately return unexported instrumentation types used only by in-package
  parity tests.

(cherry picked from commit 50b8168a2b002dca46649ab38ddbf8989e1f500c)

+++ silk/find_pred_coefs_driver_fixedpoint_libopus_parity_test.go
@@ -275,7 +275,7 @@ func TestSILKFindPredCoefsFixedLibopusParity(t *testing.T) {
 		}
 		minLag := 2 * fsKHz
 		for k := 0; k < nbSubfr; k++ {
-			in.pitchL[k] = minLag + rng.Intn(maxLag-minLag+1)
+			in.pitchL[k] = int32(minLag + rng.Intn(maxLag-minLag+1))
 		}
 
 		// res_pitch buffer: r_ptr starts at res_start; lag_ptr reaches back
@@ -283,8 +283,8 @@ func TestSILKFindPredCoefsFixedLibopusParity(t *testing.T) {
 		// res_start + (nbSubfr-1)*subfrLength + subfrLength + LTP_ORDER.
 		maxPitch := 0
 		for k := 0; k < nbSubfr; k++ {
-			if in.pitchL[k] > maxPitch {
-				maxPitch = in.pitchL[k]
+			if int(in.pitchL[k]) > maxPitch {
+				maxPitch = int(in.pitchL[k])
 			}
 		}
 		resStart := maxPitch + ltpOrder/2
